Fixes for tiff bin images

+++ segmentation/dataset.py
@@ -126,6 +126,7 @@ class MaskDataset(Dataset):
 
         mask = np.array(rescale_pil(mask, rescale_factor, 0))
         image = np.array(rescale_pil(image, rescale_factor, 1))
+
         image, mask = process(image, mask, rgb=self.rgb, preprocessing=self.preprocessing,
                               apply_preprocessing=apply_preprocessing, augmentation=self.augmentation,
                               binary_augmentation=True, color_map=self.color_map)
@@ -178,13 +179,11 @@ class XMLDataset(Dataset):
 
     def __getitem__(self, item, apply_preprocessing=True):
         image_id, mask_id = self.df.get('images')[item], self.df.get('masks')[item]
-
         image = Image.open(image_id)
         rescale_factor = get_rescale_factor(image, scale_area=self.scale_area)
 
         mask = self.mask_generator.get_mask(mask_id, rescale_factor)
-        image = np.array(rescale_pil(image, rescale_factor, 1))
-
+        image = np.array(rescale_pil(image, rescale_factor, 1), dtype="uint8")
         image, mask = process(image, mask, rgb=self.rgb, preprocessing=self.preprocessing,
                               apply_preprocessing=apply_preprocessing, augmentation=self.augmentation,
                               binary_augmentation=True, color_map=self.color_map)
